MDCRAW-3: resolve consumer issues (#9)
* adding ignore for generated static files

* initial fix for many to many mappings, integration tests for hibernate, added sql command visibility to integration tests, improved test db initializer

* MDCRAW-4: remove infinite recurrsion issue, update inline log creator for sql command interception, update test db initializer

* MDCRAW-4: add converter class to handle entity conversions

* MDCRAW-4: add initial saveOrUpdate method for article repository, clean up entities, add transactional for session executor

* MDCRAW-4: integration test for article saves

* initial naive implementation of repository layers for database models, integration tests for models

* MDCRAW-4: functional test for message decompression and consumption, renaming producer publisher to correct aws service

* cleanup and reduced debug output, added warning for container networking

* modifying the serialviews to less naive implementation

* fix serial views

+++ consumers/java/mdcrawler-consumer/src/test/java/org/perpetualnetworks/mdcrawlerconsumer/database/DataSourceTest.java
@@ -58,7 +58,7 @@ public class DataSourceTest {
         SessionFactoryStoreImpl factoryStore = new SessionFactoryStoreImpl(new DataSourceFactories(
                 new MysqlDataSourceFactory(Collections.singletonList(config))));
 
-        long testId = 8092;
+        int testId = 8092;
         final SessionFactory sessionFactory = factoryStore.getSessionFactory(Database.CRAWLER_CONSUMER);
         final Session session = sessionFactory.openSession();
         final KeywordEntity keyWordEntity = session.get(KeywordEntity.class, testId);
